chore: reduce tech debt — deduplicate queries, delete dead code, extract route logic
- Add shared SESSION_BASE_COLUMNS, getActivePlanId(), querySessionLoad() helpers
  to eliminate 7+ duplicated session column lists and repeated profile/load queries
- Delete orphaned lib/coach/conversation-memory.ts (zero imports anywhere)
- Extract ~100 lines of intensity-profile orchestration from API route into lib/
