+++ lib/training/intensity-profiles.ts
@@ -1,0 +1,122 @@
+/**
+ * Intensity profile orchestration: fetches plan sessions, computes profiles
+ * and weekly summaries, and persists them to the database.
+ */
+
+import type { SupabaseClient } from "@supabase/supabase-js";
+import {
+  computeSessionIntensityProfile,
+  computeWeeklyIntensitySummary,
+  getVisualWeight,
+  type SessionIntensityProfile
+} from "@/lib/training/intensity-profile";
+
+/**
+ * Compute and persist intensity profiles + weekly summaries for all sessions
+ * in a training plan. Returns counts of profiles and summaries written.
+ */
+export async function computeAndPersistIntensityProfiles(
+  supabase: SupabaseClient,
+  userId: string,
+  planId: string
+): Promise<{ profiles: number; summaries: number }> {
+  // Fetch all sessions for the plan
+  const { data: sessions } = await supabase
+    .from("sessions")
+    .select("id,sport,type,target,notes,duration_minutes,intent_category,date")
+    .eq("plan_id", planId)
+    .eq("user_id", userId)
+    .order("date", { ascending: true });
+
+  if (!sessions || sessions.length === 0) {
+    return { profiles: 0, summaries: 0 };
+  }
+
+  // Compute profiles
+  const rawProfiles = (sessions as Array<Record<string, unknown>>).map((s) =>
+    computeSessionIntensityProfile({
+      id: s.id as string,
+      sport: s.sport as string,
+      type: s.type as string,
+      target: (s.target as string | null) ?? null,
+      notes: (s.notes as string | null) ?? null,
+      durationMinutes: (s.duration_minutes as number) ?? 0,
+      intentCategory: (s.intent_category as string | null) ?? null
+    })
+  );
+
+  // Compute max stress for visual weight
+  const maxStress = Math.max(...rawProfiles.map((p) => p.rawStress), 1);
+  const profiles: SessionIntensityProfile[] = rawProfiles.map((p) => ({
+    ...p,
+    visualWeight: getVisualWeight(p.rawStress, maxStress)
+  }));
+
+  // Upsert profiles
+  for (const profile of profiles) {
+    await supabase
+      .from("session_intensity_profiles")
+      .upsert(
+        {
+          session_id: profile.sessionId,
+          user_id: userId,
+          primary_zone: profile.primaryZone,
+          zone_distribution: profile.zoneDistribution,
+          planned_stress_score: profile.plannedStressScore,
+          planned_duration_minutes: profile.plannedDurationMinutes,
+          stress_per_minute: profile.stressPerMinute,
+          intensity_colour: profile.intensityColour,
+          visual_weight: profile.visualWeight,
+          discipline: profile.discipline
+        },
+        { onConflict: "session_id" }
+      );
+  }
+
+  // Group by week and compute weekly summaries
+  const weekGroups = new Map<string, SessionIntensityProfile[]>();
+  for (let i = 0; i < sessions.length; i++) {
+    const session = sessions[i] as Record<string, unknown>;
+    const date = session.date as string;
+    // Find the Monday of this date's week
+    const d = new Date(`${date}T00:00:00.000Z`);
+    const day = d.getUTCDay();
+    const offset = day === 0 ? -6 : 1 - day;
+    d.setUTCDate(d.getUTCDate() + offset);
+    const weekStart = d.toISOString().slice(0, 10);
+
+    if (!weekGroups.has(weekStart)) weekGroups.set(weekStart, []);
+    weekGroups.get(weekStart)!.push(profiles[i]);
+  }
+
+  const sortedWeeks = Array.from(weekGroups.keys()).sort();
+  const summaries = [];
+  let previousSummary: { totalPlannedHours: number; totalStressScore: number } | null = null;
+
+  for (const weekStart of sortedWeeks) {
+    const weekProfiles = weekGroups.get(weekStart)!;
+    const summary = computeWeeklyIntensitySummary(weekProfiles, weekStart, previousSummary);
+    summaries.push(summary);
+    previousSummary = { totalPlannedHours: summary.totalPlannedHours, totalStressScore: summary.totalStressScore };
+
+    // Upsert weekly summary
+    await supabase
+      .from("weekly_intensity_summaries")
+      .upsert(
+        {
+          user_id: userId,
+          week_start_date: weekStart,
+          zone_distribution: summary.zoneDistribution,
+          total_planned_hours: summary.totalPlannedHours,
+          total_stress_score: summary.totalStressScore,
+          session_count: summary.sessionCount,
+          hours_delta_pct: summary.hoursDeltaPct,
+          stress_delta_pct: summary.stressDeltaPct,
+          discipline_hours: summary.disciplineHours
+        },
+        { onConflict: "user_id,week_start_date" }
+      );
+  }
+
+  return { profiles: profiles.length, summaries: summaries.length };
+}
