+++ supabase/.temp/cli-latest
@@ -1,1 +1,1 @@
-v2.84.2
+v2.90.0
